Switch output parameter values to float32

+++ samplomatic/pre_samplex/pre_samplex.py
@@ -1122,7 +1122,7 @@ class PreSamplex:
                 TensorSpecification(
                     "parameter_values",
                     (max_param_idx + 1,),
-                    np.dtype(np.float64),
+                    np.dtype(np.float32),
                     "Parameter values for the template circuit.",
                 )
             )
